Renamed MultiStreamTransformer to MutliStreamTestTransformer, moved test to integration test class and activated integration test execution for writers component by adding maven-failsafe-plugin.
