+++ components/writers/pom.xml
@@ -7,6 +7,23 @@
 	</parent>
 	<artifactId>DataCleaner-writers</artifactId>
 
+   	<build>
+		<plugins>
+			<plugin>
+				<artifactId>maven-failsafe-plugin</artifactId>
+				<version>2.19.1</version>
+				<executions>
+					<execution>
+						<goals>
+							<goal>integration-test</goal>
+							<goal>verify</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+        </plugins>
+	</build>
+
 	<dependencies>
 		<dependency>
 			<groupId>org.eobjects.datacleaner</groupId>
