+++ components/writers/src/test/java/org/datacleaner/extension/output/MultiStreamTestTransformer.java
@@ -33,7 +33,7 @@ import org.datacleaner.job.output.OutputDataStreams;
  * Test transformer which produces two output streams one containing the even rows going in and the other one
  * containing the uneven rows going in.
  */
-public class MultiStreamTransformer extends MultiStreamComponent {
+public class MultiStreamTestTransformer extends MultiStreamComponent {
     @Configured
     InputColumn<?>[] _valueColumns;
